feature: 모달 컴포넌트 정리 기능구현 #43
